refactor: split app.rs into app + headless + commands (step 3.4)
Split the 1127-line app.rs into three focused modules:

- app.rs (602 lines): Interactive REPL loop + utilities
- headless.rs (88 lines): Single-prompt headless mode
- commands.rs (486 lines): /compact, /mcp, /provider, /trust handlers

All functions use KodaAgent + KodaSession from step 3.3.
Headless mode eliminates duplicated setup code entirely.

292 tests pass. Clippy clean. Zero behavior change.
